don't prepend IN instruction
this isn't very useful (afaict, the IN instruction doesn't do much on
first boot), and can get in the way of a workflow where you plot
multiple layers as different files, which can sometimes be useful.

overall, it seems better to be explicit here - if the user wants to
initialize the plotter, they can.