add "rescale" option for faster font drawing.
this rounds the points in the output to the nearest multiple of the
given rescale value, then deduplicates points.

the time to write "Sphinx of black quartz, judge my vow." in size 200 Montserrat Regular on a HP7475A plotter, with various rescale values is shown in the following table. the text is readable at all values, but becomes noticeably pixelated at values above 10, using a Faber-Castell Pitt size "S" pen.

rescale	time (secs)
1	87
2	78
3	65
4	57
5	51
6	44
7	41
8	38
9	35
10	34
11	31
12	29
13	29
14	28
15	26
16	25
17	24
18	24
19	23
20	23